Standardize child option parsing.
Handle child args and -a args in a common way so that all children
accept the same input format.
Modify UIs to pass through setup arguments to the base class.

Instead of launching the UI directly launch the UI via the _UI subclasses
in the same way that the scripts launch the clients.